% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compas_binary}
\alias{compas_binary}
\title{COMPAS data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5278 rows and 11 columns.
}
\usage{
compas_binary
}
\description{
COMPAS data
}
\seealso{
\link[mlr3fairness:compas]{mlr3fairness::compas}
}
\keyword{datasets}
