% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3fairness.R
\name{compute_fairness}
\alias{compute_fairness}
\title{Compute fairness metrics}
\usage{
compute_fairness(data, target, prediction, protected_attribute)
}
\arguments{
\item{data}{(\code{data.table}) \cr The dataset used for predicting.}

\item{target}{(\code{character}) \cr The name of the target variable. Must be available in \code{data}.}

\item{prediction}{(\code{vector}) \cr A vector containing predictions.}

\item{protected_attribute}{(\code{character}) \cr The name(s) of the protected attributes(s). Must be available in \code{data}.}
}
\value{
a \code{double} of length 4
}
\description{
Compute fairness metrics
}
\details{
Computes our chosen fairness metrics.
}
\examples{
# Compute fairness for COMPAS data
compas <- compas_binary$data()
compute_fairness(
  data = compas,
  target = "two_year_recid",
  prediction = factor(ifelse(compas$score_text == "High", 1, 0)),
  protected_attribute = "race"
)

# Compute fairness for MLB pitch data
compute_fairness(
  data = csas25,
  target = "y",
  prediction = csas25$y_hat,
  protected_attribute = "stand"
)

# Compute fairness for Texas Ironman
# columns with non-standard data types
bad <- c("overall_time", "world_record")
ironman2 <- ironman[, !names(ironman) \%in\% bad]

compute_fairness(
  data = ironman2,
  target = "y",
  prediction = ironman$y_hat,
  protected_attribute = "gender"
)

}
\seealso{
\code{\link[mlr3fairness:compute_metrics]{mlr3fairness::compute_metrics()}}
}
