% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{csas25}
\alias{csas25}
\title{Statcast pitch data from 2024}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 346250 rows and 6 columns.
}
\source{
\url{https://statds.org/events/csas2025/challenge.html}
}
\usage{
csas25
}
\description{
Data from the CSAS 2025 Data Challenge
}
\details{
Contains data for all pitches thrown in Major League Baseball during
April, May, and June of 2024.
}
\references{
\url{https://baseballsavant.mlb.com/csv-docs}
}
\seealso{
\code{\link[baseballr:scrape_statcast_savant]{baseballr::scrape_statcast_savant()}}
}
\keyword{datasets}
