% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dipsea2021}
\alias{dipsea2021}
\title{Dipsea 2021 race results}
\format{
An object of class \code{data.frame} with 1437 rows and 14 columns.
}
\source{
\url{https://www.dipsea.org/search/dipsearesults.php}
}
\usage{
dipsea2021
}
\description{
Results for the 2021 Dipsea race.
}
\references{
\doi{10.1080/09332480.2022.2145138}
}
\keyword{datasets}
