% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick.R
\name{fairness_cv}
\alias{fairness_cv}
\alias{fairness_independence}
\alias{fairness_eod}
\alias{fairness_separation}
\alias{fairness_predictive_parity}
\alias{fairness_sufficiency}
\title{Compute fairness metrics using yardstick}
\usage{
fairness_cv(data, truth, estimate)

fairness_independence(data, truth, estimate)

fairness_eod(data, truth, estimate)

fairness_separation(data, truth, estimate)

fairness_predictive_parity(data, truth, estimate)

fairness_sufficiency(data, truth, estimate)
}
\description{
Compute fairness metrics using yardstick
}
\examples{
\dontshow{if (require(dplyr)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Compute separation measure for Ironman
ironman |>
  group_by(gender) |>
  fairness_independence(truth = y, estimate = y_hat)
\dontshow{\}) # examplesIf}
\dontshow{if (require(dplyr)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ironman |>
  group_by(gender) |>
  fairness_separation(truth = y, estimate = y_hat)
\dontshow{\}) # examplesIf}
\dontshow{if (require(dplyr)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Compute sufficiency measure for Ironman
ironman |>
  group_by(gender) |>
  fairness_sufficiency(truth = y, estimate = y_hat)
\dontshow{\}) # examplesIf}
}
