% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hof2025}
\alias{hof2025}
\title{Baseball Holl of Fame classifier data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20986 rows and 21 columns.
}
\usage{
hof2025
}
\description{
Baseball Holl of Fame classifier data
}
\references{
\url{https://baseballwithr.wordpress.com/2014/11/18/building-a-hall-of-fame-classifier-2/}
}
\keyword{datasets}
