% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ironman}
\alias{ironman}
\title{Ironman Texas race results from 2024}
\format{
\subsection{\code{ironman}}{

A data frame with 70 rows and 8 columns:
\describe{
\item{gender}{Individual's gender}
\item{division}{Individual's division within the competition}
\item{division_rank}{Individual's ranking within their division}
\item{overall_time}{Individual's total time to complete the race}
\item{finish_status}{Indicates if an individual completed the race or not}
\item{y}{Indicates if the individual was awarded prize money or not}
\item{quotient_model}{Quotient between an individual's overall time and the 2024 world reccord}
\item{y_hat}{Indicates if an individual would've recieved prize money or not using Martens and Starflinger's quotient method}
}
}
}
\source{
\url{https://www.coachcox.co.uk/imstats/race/2148/results/}
}
\usage{
ironman
}
\description{
Data from the 2024 Ironman Texas race
}
\details{
Contains data for professional racers' times and rankings as well as
predicted outcomes derived from Marten and Starflinger's proposed quotient method
}
\references{
\doi{10.1007/978-3-031-24907-5_40}
}
\keyword{datasets}
