% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fairness_matrix.R
\name{metrics_mlr3}
\alias{metrics_mlr3}
\title{Compute fairness metrics using MLR3}
\usage{
metrics_mlr3(data, truth, estimate, class)
}
\arguments{
\item{data}{The data for which to calculate the metrics}

\item{truth}{A string. The observed outcome in the data}

\item{estimate}{A string. The predicted outcome in the data}

\item{class}{A string. The protected class, a character vector with at least two levels}
}
\value{
A 1x3 matrix reporting the independence,
sufficiency, and separation scores for the data
}
\description{
\code{metrics_mlr3()} calculates the independence,
sufficiency, and separation scores for a given dataset using the \code{mlr3fairness}
package
}
\examples{
\dontshow{if (require(dplyr)) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Compute metrics for Ironman data
score_matrix <- metrics_mlr3(data = ironman, truth = "y", estimate = "y_hat", class = "gender")
\dontshow{\}) # examplesIf}
}
