% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nz_swim}
\alias{nz_swim}
\title{New Zealand master's swimming}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 13252 rows and 12 columns.
}
\source{
\url{https://osf.io/gcpwv}
}
\usage{
nz_swim
}
\description{
Time, in seconds, for New Zealand master's swimmers in different events.
The data covers years 2013-2020
}
\references{
\doi{10.1515/jqas-2023-0051}
}
\keyword{datasets}
