% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fairness}
\alias{plot_fairness}
\alias{plotly_fairness}
\title{Plot fairness measures}
\usage{
plot_fairness(data, ...)

plotly_fairness(data, ...)
}
\arguments{
\item{data}{a \link[tibble:tibble]{tibble::tibble} of output from \code{\link[=fairness_cube]{fairness_cube()}}}

\item{...}{currently ignored}
}
\description{
Plot fairness measures
}
\examples{
\dontshow{if (require(dplyr)) withAutoprint(\{ # examplesIf}
# Compute fairness measures for Ironman
d1 <- ironman |>
  group_by(gender) |>
  fairness_cube(truth = y, estimate = y_hat)

# Compute fairness measures for baseball
d2 <- csas25 |>
  group_by(stand) |>
  fairness_cube()

# Compute fairness measures for COMPAS
d3 <- compas_binary$data() |>
  mutate(is_high_risk = factor(ifelse(score_text == "High", 1, 0))) |>
  group_by(race) |>
  fairness_cube(truth = two_year_recid, estimate = is_high_risk)

d <- list(d1, d2, d3) |>
  bind_rows()

# Plot a series of 2D plots
d |>
  plot_fairness()

# Plot a 3D plots
d |>
  plotly_fairness()
\dontshow{\}) # examplesIf}
}
