% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_independence}
\alias{plot_independence}
\alias{plot_separation}
\alias{plot_sufficiency}
\title{Plot fairness metrics}
\usage{
plot_independence(data, truth = y, estimate = y_hat, ...)

plot_separation(data, truth = y, ...)

plot_sufficiency(data, truth = y, ...)
}
\arguments{
\item{data}{Either a \code{data.frame} containing the columns specified by the
\code{truth} and \code{estimate} arguments, or a \code{table}/\code{matrix} where the true
class results should be in the columns of the table.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{The column identifier for the predicted class
results (that is also \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{factor} vector.}

\item{...}{Not currently used.}
}
\description{
Plot fairness metrics
}
\examples{
\dontshow{if (require(dplyr)) withAutoprint(\{ # examplesIf}
# example code

hof_grp <- hof2025 |>
  mutate(is_pitcher = tSO > 100) |>
  group_by(is_pitcher)

hof_grp |>
  plot_independence()

hof_grp |>
  plot_separation()
\dontshow{\}) # examplesIf}
}
