% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{swimming1650}
\alias{swimming1650}
\title{Master's swimming the 1650}
\format{
An object of class \code{data.frame} with 2841 rows and 5 columns.
}
\source{
\url{https://github.com/elizabeth-upton/Age_and_Performance_SwimRun}
}
\usage{
swimming1650
}
\description{
Time, in seconds, for master's swimmers in the 1650 freestyle event.
The data covers the years 2006-2016.
}
\references{
\doi{10.1515/jqas-2024-0018}
}
\keyword{datasets}
