% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{compas_binary}
\alias{compas_binary}
\title{COMPAS data}
\format{
An object of class \code{TaskClassif} (inherits from \code{TaskSupervised}, \code{Task}, \code{R6}) of length 60.
}
\usage{
compas_binary
}
\description{
COMPAS data
}
\seealso{
\link[mlr3fairness:compas]{mlr3fairness::compas}
}
\keyword{datasets}
