% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3fairness.R
\name{compute_fairness}
\alias{compute_fairness}
\alias{compute_fairness2}
\title{Compute fairness metrics}
\usage{
compute_fairness(data, target, prediction, protected_attribute)

compute_fairness2(data, truth, estimate)
}
\arguments{
\item{data}{(\code{data.table}) \cr The dataset used for predicting.}

\item{target}{(\code{character}) \cr The name of the target variable. Must be available in \code{data}.}

\item{prediction}{(\code{vector}) \cr A vector containing predictions.}

\item{protected_attribute}{(\code{character}) \cr The name(s) of the protected attributes(s). Must be available in \code{data}.}
}
\value{
a \code{double} of length 4
}
\description{
Compute fairness metrics
}
\details{
Computes our chosen fairness metrics.
}
\examples{
# Compute fairness for COMPAS data
compas <- compas_binary$data()
compute_fairness(
  data = compas,
  target = "two_year_recid",
  prediction = factor(ifelse(compas$score_text == "High", 1, 0)),
  protected_attribute = "race"
)

# Use a tidy-interface
compas |>
  mutate(y_hat = factor(ifelse(score_text == "High", 1, 0))) |>
  group_by(race) |>
  compute_fairness2(truth = "two_year_recid", estimate = "y_hat")

# Compute fairness for MLB pitch data
compute_fairness(
  data = csas25,
  target = "y",
  prediction = csas25$y_hat,
  protected_attribute = "stand"
)

Use a tidy-interface
csas25 |>
  mutate(
    y = factor(ifelse(is_within_strike_zone, 1, 0)),
    y_hat = factor(ifelse(is_called_strike, 1, 0)),
  ) |>
  group_by(stand) |>
  compute_fairness2(
    truth = "y",
    estimate = "y_hat"
  )

# Compute fairness for Texas Ironman

# remove columns with non-standard data types
bad <- c("overall_time", "world_record")
ironman2 <- ironman[, !names(ironman) \%in\% bad]

compute_fairness(
  data = ironman2,
  target = "y",
  prediction = ironman$y_hat,
  protected_attribute = "gender"
)

}
\seealso{
\code{\link[mlr3fairness:compute_metrics]{mlr3fairness::compute_metrics()}}
}
