% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yardstick.R
\name{fairness_cv}
\alias{fairness_cv}
\alias{fairness_independence}
\alias{fairness_eod}
\alias{fairness_separation}
\alias{fairness_predictive_parity}
\alias{fairness_sufficiency}
\alias{fairness_cube}
\title{Compute fairness metrics using yardstick}
\usage{
fairness_cv(...)

fairness_independence(...)

fairness_eod(...)

fairness_separation(...)

fairness_predictive_parity(...)

fairness_sufficiency(...)

fairness_cube(data, truth = y, estimate = y_hat)
}
\arguments{
\item{...}{Not currently used.}

\item{data}{Either a \code{data.frame} containing the columns specified by the
\code{truth} and \code{estimate} arguments, or a \code{table}/\code{matrix} where the true
class results should be in the columns of the table.}

\item{truth}{The column identifier for the true class results
(that is a \code{factor}). This should be an unquoted column name although
this argument is passed by expression and supports
\link[rlang:topic-inject]{quasiquotation} (you can unquote column
names). For \verb{_vec()} functions, a \code{factor} vector.}

\item{estimate}{The column identifier for the predicted class
results (that is also \code{factor}). As with \code{truth} this can be
specified different ways but the primary method is to use an
unquoted variable name. For \verb{_vec()} functions, a \code{factor} vector.}
}
\description{
Compute fairness metrics using yardstick
}
\examples{
\dontshow{if (require(dplyr)) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\dontshow{if (require(dplyr)) withAutoprint(\{ # examplesIf}
\dontshow{\}) # examplesIf}
\dontshow{if (require(dplyr)) withAutoprint(\{ # examplesIf}
# Compute fairness measures for Ironman
ironman_grp <- ironman |>
  mutate(
    y = factor(division_rank <= 10),
    y_hat = factor(dense_rank(quotient_model) <= 20)
  ) |>
  group_by(gender)

ironman_grp |>
  fairness_cube()

ironman_grp |>
  fairness_independence()

ironman_grp |>
  fairness_separation()

ironman_grp |>
  fairness_sufficiency()

# Compute fairness measures for baseball
csas25 |>
  mutate(
    y = factor(is_within_strike_zone),
    y_hat = factor(is_called_strike)
  ) |>
  group_by(stand) |>
  fairness_cube()

# Compute fairness measures for COMPAS
compas_binary$data() |>
  mutate(is_high_risk = factor(ifelse(score_text == "High", 1, 0))) |>
  group_by(race) |>
  fairness_cube(truth = two_year_recid, estimate = is_high_risk)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[yardstick:metric_set]{yardstick::metric_set()}}, \link[yardstick:yardstick-package]{yardstick::yardstick}
}
